Fix location budgeting migration: campaign_id must be BIGINT
ads_campaigns.id is BIGINT; INT FK causes errno 150.

+++ flaskapp/migrations/add_location_based_budgeting.sql
@@ -164,7 +164,7 @@ CREATE TABLE IF NOT EXISTS campaign_location_targets (
     id INT AUTO_INCREMENT PRIMARY KEY,
 
     -- Campaign
-    campaign_id INT NOT NULL,
+    campaign_id BIGINT NOT NULL,
     account_id BIGINT NOT NULL,
 
     -- Location targeting
